refactor²
